requirements.txt
# Core AI
anthropic>=0.40.0

# Topic fetching
feedparser>=6.0.11
pytrends>=4.9.2

# Image generation
Pillow>=10.4.0

# TTS
gTTS>=2.5.3

# YouTube upload (OAuth 2.0 + Data API v3)
google-api-python-client>=2.140.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0

# HTTP (transitive, pin for stability)
requests>=2.32.0
